+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+* Changed the **format** command to update the tests field according to the test_playbooks' commands and scripts.
 * Added 'deprecated' release notes template.
 * Fixed an issue where **run-test-playbook** command failed to get the task entries when the test playbook finished with errors.
 * Fixed an issue in **validate** command when running with `no-conf-json` argument to ignore the `conf.json` file.