+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+* Added the '--use-packs-known-words' to **doc-review** command
 * Changed the **format** command to update the tests field according to the test_playbooks' commands and scripts.
 * Added 'deprecated' release notes template.
 * Fixed an issue where **run-test-playbook** command failed to get the task entries when the test playbook finished with errors.